Small changes, renaming Neurofitter finished
git-svn-id: file:///home/tclose/Desktop/neurofitter/Neurofitter@85 41e4e6bf-1111-0410-9f04-deecbdac5fd9

+++ implementation/SwarmFitterInterface.cpp
@@ -4,7 +4,7 @@
 
 FitterResults SwarmFitterInterface::runFitter(ModelTuningParameters * startPoint) {
 
-	int numberOfFlies = (int)(10.0+2.0*sqrt(toInt(fixedParams["Dimensions"])));
+	int numberOfFlies = (int)(10.0+2.0*sqrt((double)toInt(fixedParams["Dimensions"])));
 	double w = 1.0/(2.0*log(2.0));
 	double c = 0.5 + log(2.0);
